Research to create a X-scrollbar for table on small screens
More tests to improve the layout of the tables on small screens.
Used proper HTML tags for tables (instead of "divs")
Display is better as boxes of the table now remain in lines (no line break) eve when out of the screen ("display: flex" CSS property).
Still not satisfied with the scrollbar in X as several of them appear: 1 for each line (instead of 1 scrollbar for the whole table).