Fine-tuning dropdown menus for setting up events
New variables for location/duration/description choices of the event so that they are independent from the ones of the already booked event.
Initialization of the location is now always our virtual room in Google Meet.
Small updates to the style of the page (margins, etc...)
Updated the behaviour of the Save Event button so that we can save an event only when the starting time is defined. (is not active when only the date is selected)