Added an horizontal scrollbar to the list of images at the top of profile component.
Now displays 50 images accessible through horizontal scrollbar. (max 5 displayed at a time, in the space available, like before)
On computer (when the device has a fine pointer, a "mouse"), there is some more free space at the bottom of the images so that the scrollbar does not "crop" the bottom of the images.
With touchscreens, the scrollbar is far less thick and this bottom padding is not necessary.