Updated title for new chats
Now, when creating a new chat, the first title will be "General - Starting a new chat." by default. (instead of "Starting a new chat." previously)

Indeed, the title "Starting a new chat." led to the creation of a category "Starting" in the chat filters menu...which is not extremely clear and relevant.

With this new title, they will automatically be in the "General" category...up to the time when we rename them to associate them with a more accurate category.